+++ packages/plugin-app-core/src/index.ts
@@ -60,6 +60,7 @@ export default (api, options) => {
 
   // register api method
   const generator = initGenerator(api, { ...options, hasJsxRuntime });
+
   setRegisterMethod(api, { generator });
 
   // add core template for framework
@@ -134,6 +135,8 @@ function getDefaultRenderData(api, options) {
       isMPA: false,
       tabBarPath: '', // avoid ejs error
       routesFilePath: './staticConfig',
+      // MPA 下会覆盖
+      enableRouter: true,
     };
   } else {
     return {
@@ -141,6 +144,8 @@ function getDefaultRenderData(api, options) {
       isReact: true,
       isRax: false,
       ssr,
+      // MPA 下会覆盖
+      enableRouter: (!userConfig.mpa && userConfig.router !== false),
     };
   }
 }