+++ packages/plugin-router/src/runtime/_routes.ts
@@ -1,3 +1,2 @@
-// 注释：
 // 该文件仅用于 module.tsx 中引用 $ice/routes 的编译问题
 export default [];