+++ packages/plugin-react-app/CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
-## 2.0.5
+## 2.1.0
 
+- [feat] add runtime abilities: addProvider/WrapperErrorBoundary/WrapperCSR/WrapperSSR/WrapperSearchParams (move from plugin-router)
+- [feat] add runtime router options check and log warning
 - [fix] update dependencies version
 
 ## 2.0.4